requirements.txt
amqp==1.4.9
anyjson==0.3.3
apns==2.0.1
awscli==1.11.156
billiard==3.3.0.23
boto3==1.4.1
botocore==1.4.93
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
Django==1.9.2
django-celery==3.2.1
django-cors-headers==2.1.0
django-extensions==1.6.1
django-rest-swagger==0.3.5
django-seed==0.1.7
djangorestframework==3.3.2
djangorestframework-csv==2.0.0
djangorestframework-jwt==1.11.0
docopt==0.4.0
docutils==0.14
drfdocs==0.0.11
Faker==0.7.18
fuzzywuzzy==0.15.1
idna==2.6
jmespath==0.9.3
jsonfield==2.0.2
kombu==3.0.37
mandrill==1.0.57
mysqlclient==1.3.3
olefile==0.44
pbr==3.1.1
pyasn1==0.3.6
PyJWT==1.5.2
PyMySQL==0.7.9
pyotp==2.2.1
python-dateutil==2.6.1
python-gcm==0.4
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.11.1
rsa==3.4.2
s3transfer==0.1.11
six==1.10.0
stevedore==1.24.0
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper-win==1.2.1
webencodings==0.5.1
wkhtmltopdf==0.2
XlsxWriter==0.9.3
